refactor: enforce float32 renderer math

+++ simd_ray_tracer/Vec3.cpp
@@ -1,7 +1,7 @@
 #include "Vec3.h"
 
-Vec3::Vec3() : x(0), y(0), z(0) {}
-Vec3::Vec3(double v) : x(v), y(v), z(v) {}
+Vec3::Vec3() : x(0.0f), y(0.0f), z(0.0f) {}
+Vec3::Vec3(float v) : x(v), y(v), z(v) {}
 Vec3 Vec3::abs() const {
     return Vec3(std::abs(x), std::abs(y), std::abs(z));
 }
@@ -14,15 +14,15 @@ Vec3 Vec3::operator-(const Vec3& other) const {
     return Vec3(x - other.x, y - other.y, z - other.z);
 }
 
-Vec3 Vec3::operator*(double scalar) const {
+Vec3 Vec3::operator*(float scalar) const {
     return Vec3(x * scalar, y * scalar, z * scalar);
 }
 
-Vec3 Vec3::operator/(double scalar) const {
+Vec3 Vec3::operator/(float scalar) const {
     return Vec3(x / scalar, y / scalar, z / scalar);
 }
 
-double Vec3::dot(const Vec3& other) const {
+float Vec3::dot(const Vec3& other) const {
     return x * other.x + y * other.y + z * other.z;
 }
 
@@ -34,12 +34,12 @@ Vec3 Vec3::cross(const Vec3& other) const {
     );
 }
 
-double Vec3::length() const {
+float Vec3::length() const {
     return std::sqrt(x * x + y * y + z * z);
 }
 
 Vec3 Vec3::normalize() const {
-    double l = length();
+    float l = length();
     return Vec3(x / l, y / l, z / l);
 }
